fastapi/requirements.txt
fastapi
uvicorn
python-multipart
pyyaml==5.3.1
pytest-shutil
